Move 'RemoteImageDataFeedLoader' to production
